Consolidate persistent base
