+++ perst-core/src/main/java/org/garret/perst/Pinned.java
@@ -1,0 +1,6 @@
+package org.garret.perst;
+
+/**
+ * Marker interface for classes requiring pinning semantics.
+ */
+public interface Pinned {}
